% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdW.R
\name{BdW}
\alias{BdW}
\title{Beta discrete Weibull (BdW) Model for Projecting Customer Retention.}
\usage{
BdW(surv_value, h, lower = c(0.001, 0.001, 0.001), upper = c(10000,
  10000, 10000))
}
\arguments{
\item{surv_value}{a numeric vector of historical customer retention percentage should start at 100 and values should be between 0 and 100}

\item{h}{forecasting horizon}

\item{lower}{lower limit used in \code{R} \code{optim} rotuine. Default is \code{c(1e-3,1e-3)}.}

\item{upper}{upper limit used in \code{R} \code{optim} rotuine. Default is \code{c(10000,10000,10000)}.}
}
\value{
\item{fitted:}{Fitted Values based on historical data} \item{max.likelihood:}{Maximum Likelihood of Beta Geometric}
\item{params - alpha, beta c:}{Returns alpha and beta paramters from maximum likelihood estimation for beta distribution and c}
}
\description{
\code{BdW} is a beta discrete weibull model implemented based on \code{Fader and Hardie} probability based projection methedology
}
\examples{
surv_value <- c(100,86.9,74.3,65.3,59.3)
h <- 6
BdW(surv_value,h)

}
\references{
{Fader  PS and Hardie BGS (2007), How to project customer retention. Volume 21, Issue 1. Journal of Interactive Marketing.}

{Fader  PS and Hardie BGS et al. (2018), How to Project Customer Retention Revisited: The Role of Duration Dependence. Volume 43, Journal of Interactive Marketing.}
}
