% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_retention.R
\docType{data}
\name{customer_retention}
\alias{customer_retention}
\title{Observed \% Customers Surviving at Least 0-12 Years}
\format{A data frame 13 observations and 3 variables.}
\usage{
data(customer_retention)
}
\description{
A dataset containing customer retention.
}
\details{
\describe{
  \item{year}{Time in years}
  \item{regular}{\% of regular customers surviving}
  \item{high_end}{\% of high_end customers surviving}
}
}
\references{
{Fader P, Hardie B. How to project customer retention. Journal of Interactive Marketing. 2007;21(1):76-90.}
}
\keyword{datasets}
