% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_end.R
\docType{data}
\name{high_end}
\alias{high_end}
\title{\% of High End Customers Surviving Years 0 - 12.}
\format{A data frame 13 observatios:
\describe{
  \item{high_end}{\% customers high end Retained from 0 - 12 years}
}}
\usage{
high_end
}
\description{
A dataset containing cusomer retention for High End customers.
}
\references{
{Fader  PS and Hardie BGS (2007), How to project customer retention. Volume 21, Issue 1. Journal of Interactive Marketing}
}
\keyword{datasets}
