% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistency_data.R
\docType{data}
\name{persistency_data}
\alias{persistency_data}
\title{\% of patients persistency by different theraputic class.}
\format{A data frame 334 observatios and 3 variables:
\describe{
  \item{Data}{ Type of Dataset\, example Hypertension\, Epilepsy}
  \item{Time}{Time Period}
  \item{Value}{\% Patients retained by time}
}}
\usage{
data(persistency_data)
}
\description{
A dataset containing persistency of patients in different therapeutic classes.
}
\keyword{datasets}
