% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_end.R
\docType{data}
\name{regular_end}
\alias{regular_end}
\title{\% of Regular Customers Surviving Years 0 - 12.}
\format{A data frame 13 observatios:
\describe{
  \item{regular_end}{\% customers regular end Retained from 0 - 12 years}
}}
\usage{
regular_end
}
\description{
A dataset containing cusomer retention for Regular customers
}
\references{
{Fader  PS and Hardie BGS (2007), How to project customer retention. Volume 21, Issue 1. Journal of Interactive Marketing}
}
\keyword{datasets}
