% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_sbg.R
\name{shiftedBetaGeometric}
\alias{shiftedBetaGeometric}
\title{Shifted Beta-geometric (sbg) Distribution Family Function}
\usage{
dsbg(x, shape1, shape2, log = FALSE)
psbg(x, shape1, shape2, lower.tail = TRUE, log.p = FALSE)
qsbg(p, shape1, shape2, lower.tail = TRUE, log.p = FALSE)
rsbg(n, shape1, shape2)
}
\arguments{
\item{x}{Vector of non-negative integers for \code{dsbg} and \code{psbg}.}

\item{p}{Vector of probabilities (0 <= p <= 1) for \code{qsbg}.}

\item{n}{Number of random variates to generate for \code{rsbg}.}

\item{shape1}{First shape parameter "a" (must be > 0).}

\item{shape2}{Second shape parameter "b" (must be > 0).}

\item{log}{Logical; if TRUE, probabilities are returned on the log scale (for \code{dsbg} and \code{psbg}).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P(X <= x), otherwise P(X > x) (for \code{psbg}).}

\item{log.p}{Logical; if TRUE, probabilities are returned on the log scale (for \code{psbg} and \code{qsbg}).}
}
\value{
\itemize{
\item \code{dsbg}: A numeric vector of PMF values.
\item \code{psbg}: A numeric vector of CDF values.
\item \code{qsbg}: A numeric vector of quantile values.
\item \code{rsbg}: A numeric vector of random variates.
}
}
\description{
Provides functions for the probability mass function (PMF), cumulative distribution function (CDF), quantile function, and random variate generation for the SBG distribution.
}
\details{
The Shifted Beta Geometric distribution with two shape parameters shape1 (\eqn{a}) and shape2 (\eqn{b}) has the following CDF:
\deqn{1- Beta(a,b+x)/Beta(a,b)}

For \eqn{x= 1,2,3,...,n} and \eqn{a > 0} and \eqn{b > 0}.

The Shifted Beta Geometric (sBG) distribution, is a probability mixutre model of Beta and Geometric distributions. sBG was introduced by Fader and Hardie, models customer retention by assuming that
individuals have heterogeneous dropout probabilities. These probabilities are drawn from a Beta distribution, and each customer's retention follows a geometric process. This combination captures variability in churn behavior across a population,
making it well-suited for analyzing survival data, customer lifetime and retention data.
}
\examples{
# PMF example
dsbg(1:5, shape1 = 2, shape2 = 3)

# CDF example
psbg(1:5, shape1 = 2, shape2 = 3)

# Quantile example
qsbg(c(0.1, 0.5, 0.9), shape1 = 2, shape2 = 3)

# Random variates
rsbg(10, shape1 = 2, shape2 = 3)

}
\references{
{Fader P, Hardie B. How to project customer retention. Journal of Interactive Marketing. 2007;21(1):76-90.}

{Fader P, Hardie B, Liu Y, Davin J, Steenburgh T. "How to Project Customer Retention" Revisited: The Role of Duration Dependence. Journal of Interactive Marketing. 2018;43:1-16.}
}
